add error handling in UpdateFormByBorrow in BorrowManager

+++ BorrowManager.gs
@@ -150,9 +150,49 @@ function ResisterStatus(answers){
   }
 }
 
-function UpdateFormByBorrow(answers) {
+function UpdateFormByBorrow(answers){
+  // var answers = {
+  //   "bookNumber" : 1,
+  //   "employeeName": "山田太郎",
+  //   "employeeNumber": 1111,
+  //   "borrowDate": new Date,
+  //   "backDeadline": new Date
+  // };//TODO:配列から取ってくる
+
+  try {
+    const SS = SpreadsheetApp.openById("1d-DK2eNTH6iUVlj_kyNE6lvSp20eQiIR1ydu-6lf9RA");
+  }
+  catch (e) {
+    // Logger.log("error");
+    let error = {};
+    error.timestamp = new Date(),"JST", "yyyy/MM/dd HH:mm:ss";
+    error.book = answers.bookNumber +"-貸出";
+    error.employeeName = answers.employeeName;
+    error.employeeNumber = answers.employeeNumber;
+    error.formAnswer1 = answers.borrowDate;
+    error.formAnswer2 = answers.backDeadline;
+    error.where = "UpdateFormByBorrow(BorrowManager)";
+    error.what = "スプレッドシート「図書貸出管理」のIDが間違っています";
+    InsertError(error);
+    return;
+  }
   const SS = SpreadsheetApp.openById("1d-DK2eNTH6iUVlj_kyNE6lvSp20eQiIR1ydu-6lf9RA");
+
   const STATUS_SHEET = SS.getSheetByName("貸出状況");
+  // Logger.log(STATUS_SHEET);
+  if (STATUS_SHEET == null || STATUS_SHEET == ""){
+    let error = {};
+    error.timestamp = new Date(),"JST", "yyyy/MM/dd HH:mm:ss";
+    error.book = answers.bookNumber;
+    error.employeeName = answers.employeeName;
+    error.employeeNumber = answers.employeeNumber;
+    error.formAnswer1 = answers.borrowDate;
+    error.formAnswer2 = answers.backDeadline;
+    error.where = "UpdateFormByBorrow(BorrowManager)";
+    error.what = "スプレッドシート「図書貸出管理」内，「貸出状況」シートの名前が間違っています";
+    InsertError(error);
+    return;
+  }
   let range = STATUS_SHEET.getRange("A:G");
   let lastRow = STATUS_SHEET.getLastRow();
 
@@ -161,24 +201,91 @@ function UpdateFormByBorrow(answers) {
   //  Logger.log(answers.backDeadline);
   // answers.backDeadline.toString();
   // Logger.log(answers.backDeadline);
+  if (answers.bookNumber == null || answers.bookNumber == "" 
+      || answers.backDeadline == null || answers.backDeadline ==""){
+    let error = {};
+    error.timestamp = new Date(),"JST", "yyyy/MM/dd HH:mm:ss";
+    error.book = answers.bookNumber + "-貸出";
+    error.employeeName = answers.employeeName;
+    error.employeeNumber = answers.employeeNumber;
+    error.formAnswer1 = answers.borrowDate;
+    error.formAnswer2 = answers.backDeadline;
+    error.where = "UpdateFormByBorrow(BorrowManager)";
+    error.what = "answersが取得できませんでした";
+    InsertError(error);
+    return;
+  }
   answers.backDeadline = Utilities.formatDate(answers.backDeadline,"JST", "yyyy/MM/dd");
 
+  let flag = 0;
   for (let i = 2; i <= lastRow; i++){
     if (range.getCell(i, 1).getValue() == answers.bookNumber){
+      if (flag > 0){
+        let error = {};
+        error.timestamp = new Date(),"JST", "yyyy/MM/dd HH:mm:ss";
+        error.book = answers.bookNumber + "-貸出";
+        error.employeeName = answers.employeeName;
+        error.employeeNumber = answers.employeeNumber;
+        error.formAnswer1 = answers.borrowDate;
+        error.formAnswer2 = answers.backDeadline;
+        error.where = "UpdateFormByBorrow(BorrowManager)";
+        error.what = "「貸出状況」シートから書籍番号" + answers.bookNumber + "が２か所以上見つかりました";
+        InsertError(error);
+        return;
+      }
       var formId = range.getCell(i, 7).getValue();
-      //TODO:ひとつもないorふたつ以上あったらエラー
+      flag++
     }
+  }
+  if (flag == 0){
+    let error = {};
+    error.timestamp = new Date(),"JST", "yyyy/MM/dd HH:mm:ss";
+    error.book = answers.bookNumber + "-貸出";
+    error.employeeName = answers.employeeName;
+    error.employeeNumber = answers.employeeNumber;
+    error.formAnswer1 = answers.borrowDate;
+    error.formAnswer2 = answers.backDeadline;
+    error.where = "UpdateFormByBorrow(BorrowManager)";
+    error.what = "「貸出状況」シートから書籍番号が見つかりませんでした";
+    InsertError(error);
+    return;
+  }
+  if (formId == null || formId == ""){
+    let error = {};
+    error.timestamp = new Date(),"JST", "yyyy/MM/dd HH:mm:ss";
+    error.book = answers.bookNumber + "-貸出";
+    error.employeeName = answers.employeeName;
+    error.employeeNumber = answers.employeeNumber;
+    error.formAnswer1 = answers.borrowDate;
+    error.formAnswer2 = answers.backDeadline;
+    error.where = "UpdateFormByBorrow(BorrowManager)";
+    error.what = "「貸出状況」シートにフォームIDがありません";
+    InsertError(error);
+    return;
   }
   
   // var formId = statusRange.getCell(BookNumber + 1, 7).getValue();
 
   var form = FormApp.openById(formId);
+  if (form == null || form == ""){
+    let error = {};
+    error.timestamp = new Date(),"JST", "yyyy/MM/dd HH:mm:ss";
+    error.book = answers.bookNumber + "-貸出";
+    error.employeeName = answers.employeeName;
+    error.employeeNumber = answers.employeeNumber;
+    error.formAnswer1 = answers.borrowDate;
+    error.formAnswer2 = answers.backDeadline;
+    error.where = "UpdateFormByBorrow(BorrowManager)";
+    error.what = "「貸出状況」シートのフォームIDが間違っています";
+    InsertError(error);
+    return;
+  }
  
   let items = form.getItems();  
   for (let i = 0; i < items.length; i++){
     form.deleteItem(items[i]);
   }
-  form.setDescription("貸出中につき現在借りられません。しばらくお待ちください。 \n 返却予定日：" + answers.backDeadline);
+  form.setDescription("貸出中につき現在借りられません。しばらくお待ちください。 \n返却予定日：" + answers.backDeadline);
 }
